--- Frameworks/OMSDK_Prebidorg.framework/Modules/module.modulemap
@@ -1,4 +1,0 @@
-module OMSDK_Prebidorg {
-    header "Headers/OMIDImports.h"
-    export *
-}
