--- Frameworks/OMSDK_Prebidorg.framework/OMSDK_Prebidorg
@@ -1,1 +1,0 @@
-Versions/Current/OMSDK_Prebidorg
