+++ Frameworks/OMSDK_Prebidorg.xcframework/ios-arm64_armv7/OMSDK_Prebidorg.framework/Modules/module.modulemap
@@ -1,0 +1,4 @@
+module OMSDK_Prebidorg {
+    header "Headers/OMIDImports.h"
+    export *
+}
