+++ Frameworks/OMSDK_Prebidorg.xcframework/ios-arm64_i386_x86_64-simulator/OMSDK_Prebidorg.framework/Modules/module.modulemap
@@ -1,0 +1,4 @@
+module OMSDK_Prebidorg {
+    header "Headers/OMIDImports.h"
+    export *
+}
