+++ Frameworks/OMSDK_Prebidorg.xcframework/ios-arm64_i386_x86_64-simulator/OMSDK_Prebidorg.framework/_CodeSignature/CodeResources
@@ -1,0 +1,312 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>files</key>
+	<dict>
+		<key>Headers/OMIDAdEvents.h</key>
+		<data>
+		WL+i0CVe4DyR75hGVHCS8ntS8bI=
+		</data>
+		<key>Headers/OMIDAdSession.h</key>
+		<data>
+		fcrZjS8K1XowIeSNhZTzpQU8W8w=
+		</data>
+		<key>Headers/OMIDAdSessionConfiguration.h</key>
+		<data>
+		zn2xHb3hkrFtt8NjmLcahFfru7o=
+		</data>
+		<key>Headers/OMIDAdSessionContext.h</key>
+		<data>
+		K4q8gQpt/HD1ZRnxoqbtaT2dhog=
+		</data>
+		<key>Headers/OMIDFriendlyObstructionType.h</key>
+		<data>
+		ymPA8TI4PL8pK0zhc5ilyQj/SsA=
+		</data>
+		<key>Headers/OMIDImports.h</key>
+		<data>
+		kklBQQiiHyi6ArUe1OOfy92xxX4=
+		</data>
+		<key>Headers/OMIDMediaEvents.h</key>
+		<data>
+		6vBBDJ3kh4miSsNkZXDl/kCRqN4=
+		</data>
+		<key>Headers/OMIDPartner.h</key>
+		<data>
+		H5bAR3fn1dPaMB5/Db7kTp8FEwY=
+		</data>
+		<key>Headers/OMIDSDK.h</key>
+		<data>
+		iTTj30pct8qnwmyv0c2NCIrJK6g=
+		</data>
+		<key>Headers/OMIDScriptInjector.h</key>
+		<data>
+		FIriwfmX3CRcMaQhGdJZchgT54c=
+		</data>
+		<key>Headers/OMIDVASTProperties.h</key>
+		<data>
+		cmPCccS1OEvjy+33MaiW7St0Z3s=
+		</data>
+		<key>Headers/OMIDVerificationScriptResource.h</key>
+		<data>
+		cbIpjYE6zhEsN3RKuOqawxSsd3E=
+		</data>
+		<key>Headers/OMSDK.h</key>
+		<data>
+		SwK6aatAfFlLHnP2+J8eYoZyaqs=
+		</data>
+		<key>Info.plist</key>
+		<data>
+		2vKXMoBLz21/J43Vrk917sKHOaY=
+		</data>
+		<key>Modules/module.modulemap</key>
+		<data>
+		lZzWCqJnb4iFhevA4Oj89tMj2mU=
+		</data>
+	</dict>
+	<key>files2</key>
+	<dict>
+		<key>Headers/OMIDAdEvents.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			WL+i0CVe4DyR75hGVHCS8ntS8bI=
+			</data>
+			<key>hash2</key>
+			<data>
+			vkHFqUD8aSNzQAv8pI0DKg+Z5yBmggRaDoJepW3Py5A=
+			</data>
+		</dict>
+		<key>Headers/OMIDAdSession.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			fcrZjS8K1XowIeSNhZTzpQU8W8w=
+			</data>
+			<key>hash2</key>
+			<data>
+			/BQ8Xidal5ht43an8bHHO3jexkRF5FOPLwfno439b7I=
+			</data>
+		</dict>
+		<key>Headers/OMIDAdSessionConfiguration.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			zn2xHb3hkrFtt8NjmLcahFfru7o=
+			</data>
+			<key>hash2</key>
+			<data>
+			xU+4IUgfyZnIJ3UCP5dmWZ+M+w/Zchzr/eWvVTE3Iek=
+			</data>
+		</dict>
+		<key>Headers/OMIDAdSessionContext.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			K4q8gQpt/HD1ZRnxoqbtaT2dhog=
+			</data>
+			<key>hash2</key>
+			<data>
+			IVcpCQr7h/AfHCneHZLLPMccki64QaBBtGZBw94AMCQ=
+			</data>
+		</dict>
+		<key>Headers/OMIDFriendlyObstructionType.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			ymPA8TI4PL8pK0zhc5ilyQj/SsA=
+			</data>
+			<key>hash2</key>
+			<data>
+			/ZmBFlTK7mS0fZy4WqxCTpfhwDAEA4qlDHoed4FawBg=
+			</data>
+		</dict>
+		<key>Headers/OMIDImports.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			kklBQQiiHyi6ArUe1OOfy92xxX4=
+			</data>
+			<key>hash2</key>
+			<data>
+			WWfw+Z3RNrtOfWYdio9jyxKjmeVbPXzWNH7Pa7ErD0U=
+			</data>
+		</dict>
+		<key>Headers/OMIDMediaEvents.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			6vBBDJ3kh4miSsNkZXDl/kCRqN4=
+			</data>
+			<key>hash2</key>
+			<data>
+			S25KyasS4MdytLvoFfdLG3vLT/CcGYxGS5xXnngcx+0=
+			</data>
+		</dict>
+		<key>Headers/OMIDPartner.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			H5bAR3fn1dPaMB5/Db7kTp8FEwY=
+			</data>
+			<key>hash2</key>
+			<data>
+			u2u77husIJvmuQWG0lwDpkLFdN3jaP6qM4O5N8mMo/Q=
+			</data>
+		</dict>
+		<key>Headers/OMIDSDK.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			iTTj30pct8qnwmyv0c2NCIrJK6g=
+			</data>
+			<key>hash2</key>
+			<data>
+			6vHNykyxCTUsFYIRpzoxG2TBQT0u9tdpGRppj0s8Flw=
+			</data>
+		</dict>
+		<key>Headers/OMIDScriptInjector.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			FIriwfmX3CRcMaQhGdJZchgT54c=
+			</data>
+			<key>hash2</key>
+			<data>
+			fcguujztejv7pUWmtPH1JBsMKrCBxZoedW5ds7xFauM=
+			</data>
+		</dict>
+		<key>Headers/OMIDVASTProperties.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			cmPCccS1OEvjy+33MaiW7St0Z3s=
+			</data>
+			<key>hash2</key>
+			<data>
+			vbN9YCvrKcrSo7h/nO+IqM/bZECxgLd7zfxyZvHSZhI=
+			</data>
+		</dict>
+		<key>Headers/OMIDVerificationScriptResource.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			cbIpjYE6zhEsN3RKuOqawxSsd3E=
+			</data>
+			<key>hash2</key>
+			<data>
+			XQZCkq7zWn2tKhxCAIntEo2ZQ178TvRdbO5mcIDpEQ0=
+			</data>
+		</dict>
+		<key>Headers/OMSDK.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			SwK6aatAfFlLHnP2+J8eYoZyaqs=
+			</data>
+			<key>hash2</key>
+			<data>
+			RMvQNEJLOImLem6a8DtHWs08tFMncdsWXtF/QfYUIvQ=
+			</data>
+		</dict>
+		<key>Modules/module.modulemap</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			lZzWCqJnb4iFhevA4Oj89tMj2mU=
+			</data>
+			<key>hash2</key>
+			<data>
+			2EO2mqkiP/1xjI2y++ibMHIbc2PqEpgcJVvsPSLjdnQ=
+			</data>
+		</dict>
+	</dict>
+	<key>rules</key>
+	<dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^version.plist$</key>
+		<true/>
+	</dict>
+	<key>rules2</key>
+	<dict>
+		<key>.*\.dSYM($|/)</key>
+		<dict>
+			<key>weight</key>
+			<real>11</real>
+		</dict>
+		<key>^(.*/)?\.DS_Store$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>2000</real>
+		</dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^Info\.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^PkgInfo$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^embedded\.provisionprofile$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^version\.plist$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+	</dict>
+</dict>
+</plist>
